rename old and new components
