+++ WebUI/src/components/SettingsImageEdit.vue
