+++ WebUI/src/components/SettingsImageGen.vue
