+++ WebUI/src/components/SettingsPanel.vue
