+++ WebUI/src/components/SettingsVideo.vue
